+++ CHANGELOG.md
@@ -12,6 +12,7 @@ No unreleased changes :fire: -->
 
 ### :warning: BREAKING CHANGE :warning:
  - Rename carousel list props returned from `useCarousel` from `scrollContainer` to `list`
+ - Rename `onLayoutUpdate` to `onLayoutChange` (in `useTargetZone` and `usePages`)
 
 ### :rocket: Added
  - Add `lwc-hide` class name to `CarouselArrow` when it is being hidden